+++ framework/languages/dotnet/src/Zlink.Framework/Runtime/Backend/DotNet/Wrappers/ZLinkBackendDealerSocketWrapper.cs
@@ -27,12 +27,12 @@ internal sealed class ZLinkBackendDealerSocketWrapper(IDealerSocket nativeSocket
         nativeSocket.Options.MaxMessageSize = value;
     }
 
-    public void SetSendHighWaterMark(int value)
+    public void SetSendHighWaterMark(ulong value)
     {
         nativeSocket.Options.SendHighWaterMark = value;
     }
 
-    public void SetReceiveHighWaterMark(int value)
+    public void SetReceiveHighWaterMark(ulong value)
     {
         nativeSocket.Options.ReceiveHighWaterMark = value;
     }
