+++ framework/languages/dotnet/src/Zlink.Framework/Runtime/Backend/DotNet/Wrappers/ZLinkBackendSubscriberSocketWrapper.cs
@@ -21,12 +21,12 @@ internal sealed class ZLinkBackendSubscriberSocketWrapper(ISubSocket nativeSocke
         nativeSocket.Options.MaxMessageSize = value;
     }
 
-    public void SetSendHighWaterMark(int value)
+    public void SetSendHighWaterMark(ulong value)
     {
         nativeSocket.Options.SendHighWaterMark = value;
     }
 
-    public void SetReceiveHighWaterMark(int value)
+    public void SetReceiveHighWaterMark(ulong value)
     {
         nativeSocket.Options.ReceiveHighWaterMark = value;
     }
